Added PCF8563 module
Adafruit.io time server sometimes has a hiccup that confuses the Magtag. Added a PCF8563 module via the Stemma QT port to keep track of time locally, powered off the Magtag's battery. The lunch menu URL also returns a current timestamp, so it will use that to periodically sync. If power goes out, the PCF8563 defaults back to 2000-01-01, so the Magtag will now reach out to Adafruit.io for a resync if the time goes back to the default value.